Recup personnages

+++ FRoG-Creator.Serveur/Modules/ModVar.vb
@@ -24,6 +24,6 @@ Module ModVar
 
     ' -- Variables Max --
     Public MAX_PLAYERS As Integer = 500
-    Public MAX_CHARS As Short = 3
+    Public MAX_CHARS As Byte = 3
 
 End Module
